moving base classes for auth and topic plugins into common file
