+++ amqtt/plugins/base.py
@@ -1,6 +1,8 @@
 from typing import Any, Generic, TypeVar
 
+from amqtt.broker import Action
 from amqtt.plugins.manager import BaseContext
+from amqtt.session import Session
 
 C = TypeVar("C", bound=BaseContext)
 
@@ -24,3 +26,62 @@ class BasePlugin(Generic[C]):
 
     async def close(self) -> None:
         """Override if plugin needs to clean up resources upon shutdown."""
+
+
+class BaseTopicPlugin(BasePlugin[BaseContext]):
+    """Base class for topic plugins."""
+
+    def __init__(self, context: BaseContext) -> None:
+        super().__init__(context)
+
+        self.topic_config: dict[str, Any] | None = self._get_config_section("topic-check")
+        if self.topic_config is None:
+            self.context.logger.warning("'topic-check' section not found in context configuration")
+
+    async def topic_filtering(
+        self, *, session: Session | None = None, topic: str | None = None, action: Action | None = None
+    ) -> bool:
+        """Logic for filtering out topics.
+
+        Args:
+            session: amqtt.session.Session
+            topic: str
+            action: amqtt.broker.Action
+
+        Returns:
+            bool: `True` if topic is allowed, `False` otherwise
+
+        """
+        if not self.topic_config:
+            # auth config section not found
+            self.context.logger.warning("'topic-check' section not found in context configuration")
+            return False
+        return True
+
+
+class BaseAuthPlugin(BasePlugin[BaseContext]):
+    """Base class for authentication plugins."""
+
+    def __init__(self, context: BaseContext) -> None:
+        super().__init__(context)
+
+        self.auth_config: dict[str, Any] | None = self._get_config_section("auth")
+        if not self.auth_config:
+            self.context.logger.warning("'auth' section not found in context configuration")
+
+    async def authenticate(self, *, session: Session) -> bool | None:
+        """Logic for session authentication.
+
+        Args:
+            session: amqtt.session.Session
+
+        Returns:
+            - `True` if user is authentication succeed, `False` if user authentication fails
+            - `None` if authentication can't be achieved (then plugin result is then ignored)
+
+        """
+        if not self.auth_config:
+            # auth config section not found
+            self.context.logger.warning("'auth' section not found in context configuration")
+            return False
+        return True
