+++ tests/plugins/test_manager.py
@@ -4,9 +4,8 @@ import unittest
 
 from amqtt.broker import Action
 from amqtt.events import BrokerEvents
-from amqtt.plugins.authentication import BaseAuthPlugin
 from amqtt.plugins.manager import BaseContext, PluginManager
-from amqtt.plugins.topic_checking import BaseTopicPlugin
+from amqtt.plugins.base import BaseTopicPlugin, BaseAuthPlugin
 from amqtt.session import Session
 
 formatter = "[%(asctime)s] %(name)s {%(filename)s:%(lineno)d} %(levelname)s - %(message)s"
